requirements.txt
# Production Dependencies
# Core functionality
supabase>=2.0.0,<3.0.0
requests>=2.31.0,<3.0.0
python-frontmatter>=1.0.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
markdown>=3.5.0,<4.0.0

# AI APIs
google-generativeai>=0.3.0,<1.0.0
openai>=1.0.0,<2.0.0
anthropic>=0.8.0,<1.0.0

# Web & UI
flask>=3.0.0,<4.0.0
textual>=0.47.0,<1.0.0

# macOS App Building (optional)
pywebview>=4.0.0,<5.0.0
py2app>=0.28.0,<1.0.0
